% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{Account.bidderLocation}
\alias{Account.bidderLocation}
\title{Account.bidderLocation Object}
\usage{
Account.bidderLocation()
}
\value{
Account.bidderLocation object
}
\description{
Account.bidderLocation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Account functions: \code{\link{Account}},
  \code{\link{accounts.patch}},
  \code{\link{accounts.update}}
}

