% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{Account}
\alias{Account}
\title{Ad Exchange Buyer API Objects 
Accesses your bidding-account information, submits creatives for validation, finds available direct deals, and retrieves performance reports.}
\usage{
Account(Account.bidderLocation = NULL, bidderLocation = NULL,
  cookieMatchingNid = NULL, cookieMatchingUrl = NULL, id = NULL,
  maximumActiveCreatives = NULL, maximumTotalQps = NULL,
  numberActiveCreatives = NULL)
}
\arguments{
\item{Account.bidderLocation}{The \link{Account.bidderLocation} object or list of objects}

\item{bidderLocation}{Your bidder locations that have distinct URLs}

\item{cookieMatchingNid}{The nid parameter value used in cookie match requests}

\item{cookieMatchingUrl}{The base URL used in cookie match requests}

\item{id}{Account id}

\item{maximumActiveCreatives}{The maximum number of active creatives that an account can have, where a creative is active if it was inserted or bid with in the last 30 days}

\item{maximumTotalQps}{The sum of all bidderLocation}

\item{numberActiveCreatives}{The number of creatives that this account inserted or bid with in the last 30 days}
}
\value{
Account object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 22:30:24
filename: /Users/mark/dev/R/autoGoogleAPI/googleadexchangebuyerv12.auto/R/adexchangebuyer_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Account Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration data for an Ad Exchange buyer account.
}
\seealso{
Other Account functions: \code{\link{Account.bidderLocation}},
  \code{\link{accounts.patch}},
  \code{\link{accounts.update}}
}

