% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{AccountsList}
\alias{AccountsList}
\title{AccountsList Object}
\usage{
AccountsList(items = NULL)
}
\arguments{
\item{items}{A list of accounts}
}
\value{
AccountsList object
}
\description{
AccountsList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An account feed lists Ad Exchange buyer accounts that the user has access to. Each entry in the feed corresponds to a single buyer account.
}

