% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_functions.R
\name{creatives.list}
\alias{creatives.list}
\title{Retrieves a list of the authenticated user's active creatives. A creative will be available 30-40 minutes after submission.}
\usage{
creatives.list(maxResults = NULL, pageToken = NULL, statusFilter = NULL)
}
\arguments{
\item{maxResults}{Maximum number of entries returned on one result page}

\item{pageToken}{A continuation token, used to page through ad clients}

\item{statusFilter}{When specified, only creatives having the given status are returned}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/adexchange.buyer
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/adexchange.buyer)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/ad-exchange/buyer-rest}{Google Documentation}
}
